Add opt_merge_wires pass
Rearange connections from/to weakly connected components of wires to
always be to a single wire of the component. Useful for designs with
high-impedance signals.

From smaller tests opt_clean appears to do the same thing this pass is
doing, but it was not working to the extend I need in the dmgcpu
project.