Fix consteval, when cell evaluate more than one signal
When a signal is evaluated, all its driver cells are evaluated, and they
results are resolved to a final const value of the signal. A signal is
said to be already evaluated if it contains a const value. But if any of
the driver cells is also driving another signal, that signal will be
marked as evaluated, even if some of its driver cells have not.

To fix that, we now use a separated `pool` to mark evaluated signals.